added a search for contact function
Added the ability to search for a user by typing in the text box. the
name will be displayed in the user edit boxes so that it can be edited
and updated.
In order to load a new page, the information has to be sent back to the
controller who should then load the page with the data of that contact